api/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.2
requests==2.26.0
gunicorn==20.1.0
pydantic==1.9.0
